Move to codestyle
other 'unnecessary*' rule are there
