+++ pmd-java/src/main/resources/category/java/bestpractices.xml
@@ -1358,37 +1358,6 @@ class Foo{
         </example>
     </rule>
 
-
-    <rule name="UnnecessaryConversion"
-          language="java"
-          since="7.0.0"
-          message="Unnecessary {0}"
-          class="net.sourceforge.pmd.lang.java.rule.bestpractices.UnnecessaryConversionRule"
-          externalInfoUrl="${pmd.website.baseurl}/pmd_rules_java_bestpractices.html#unnecessaryboxing">
-        <description>
-            Reports explicit conversions that may safely be removed, either because
-            they would implicitly take place (eg widening a `byte` to an `int`), or
-            because they're semantically a noop (eg unboxing a value to rebox it immediately).
-
-            This handles widening conversion between primitives, boxing and unboxing
-            conversions (which since Java 5 are inserted by the compiler implicitly),
-            and unnecessary casts.
-        </description>
-        <priority>3</priority>
-        <example><![CDATA[
-            {
-                // Instead of
-                Integer integer = Integer.valueOf(2);
-                // you may just write
-                Integer integer = 2;
-
-                int i = integer.intValue(); // similarly for unboxing
-                // todo cast examples
-            }
-        ]]>
-        </example>
-    </rule>
-
     <rule name="UnusedAssignment"
           language="java"
           since="6.26.0"
